plant-protective-ml-main/requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart 
pillow 
python-dotenv  
